+++ unix.c
@@ -56,9 +56,9 @@ struct mill_unixlistener {
 struct mill_unixconn {
     struct mill_unixsock sock;
     int fd;
-    int ifirst;
-    int ilen;
-    int olen;
+    size_t ifirst;
+    size_t ilen;
+    size_t olen;
     char ibuf[MILL_UNIX_BUFLEN];
     char obuf[MILL_UNIX_BUFLEN];
 };
@@ -348,7 +348,7 @@ size_t unixrecv(unixsock s, void *buf, size_t len, int64_t deadline) {
                     return len - remaining;
                 sz = 0;
             }
-            if(sz == remaining) {
+            if((size_t)sz == remaining) {
                 errno = 0;
                 return len;
             }
@@ -368,7 +368,7 @@ size_t unixrecv(unixsock s, void *buf, size_t len, int64_t deadline) {
                     return len - remaining;
                 sz = 0;
             }
-            if(sz < remaining) {
+            if((size_t)sz < remaining) {
                 memcpy(pos, conn->ibuf, sz);
                 pos += sz;
                 remaining -= sz;
